Move models + rework
