--- backend/src/domain/participant.py
@@ -1,13 +1,0 @@
-from dataclasses import dataclass
-
-from backend.src.domain.meter import Meter
-
-
-@dataclass
-class Participant:
-
-    active: bool
-    meters: list[Meter]
-
-    def add_meterdata(self, meterdata):
-        self.meterdata.append(meterdata)
