+++ backend/src/models/community.py
@@ -1,0 +1,10 @@
+import dataclasses
+
+from models.participant import Participant
+
+CommunityId=int
+
+@dataclasses.dataclass
+class Community:
+    id: CommunityId
+    participants: list[Participant]
