+++ backend/src/models/participant.py
@@ -1,0 +1,16 @@
+from dataclasses import dataclass
+
+from models.meter import Meter
+
+ParticipantId=int
+
+@dataclass
+class Participant:
+    active: bool
+    meters: list[Meter]
+
+
+@dataclass
+class ParticipantResult:
+    participantId: ParticipantId
+    value: int
