+++ backend/src/models/settlement.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+from models.community import CommunityId
+from models.participant import ParticipantResult
+
+
+@dataclass
+class SettlementResult:
+    communityId: CommunityId
+    results: list[ParticipantResult]
